% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\name{extractTrawlsBiotic}
\alias{extractTrawlsBiotic}
\title{Extract Trawl}
\usage{
extractTrawlsBiotic(
  biotic,
  serialnumbers = NULL,
  sampleQuality = NULL,
  gearCodes = NULL
)
}
\arguments{
\item{biotic}{NMDbiotic data (v3.x) as parsed by RstoxData::readXmlFile}

\item{serialnumbers}{gear codes to include. If NULL all are included. Compared to column (fishstation/gear)}

\item{sampleQuality}{sample quality codes to include. If NULL all are included. Compared to column (fishstation/samplequality)}
}
\value{
\code{\link[acousticsQA]{trawlLocation}}
}
\description{
Extract trawl locations from Biotic data
}
