% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\name{horisontalProfile}
\alias{horisontalProfile}
\title{Horisontal Profile}
\description{
sa in horisontal bands of the water column, along with their distance from seadbed (bottom)

sa in vertical bands of the water column, along with log (sailed distance),
 position parameters and time parameters
}
\details{
data table with columns:
 \describe{
  \item{height}{Factor for each horisontal band. Names indicate interval in meters.}
  \item{sa}{Aggregated sa in horsintal band}
 }

data table with columns:
 \describe{
  \item{sa}{Aggregated sa in vertical band}
  \item{log}{log, sailed distance in nmi}
  \item{time}{time as POSIXct}
  \item{latitude}{latitude in decimal degrees, WGS84}
  \item{longitude}{longitude in decimal degrees, WGS84}
  \item{distance}{width of vertical band in nmi}
 }
}
