% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\name{horisontalProfileLUF20}
\alias{horisontalProfileLUF20}
\title{Extract Horisontal profile}
\usage{
horisontalProfileLUF20(echosounder, targetAcocat, bins = 20, freq = 38000)
}
\arguments{
\item{echosounder}{luf20 as read by RstoxData::readXmlFile}

\item{targetAcocat}{acoustic category to extract profile for}

\item{bins}{number of bins for the profile}

\item{freq}{frequncy of transreceiver to extract data from.}
}
\value{
\code{\link[acousticsQA]{horisontalProfile}}
}
\description{
Makes horisontal profile (sa by approximate distance from bottom) from pelagic channels
}
\details{
Consult reference table for acoustic categories.
 https://referenceeditor.hi.no/apps/referenceeditor/v2/tables/acousticCategory

 some common options:
 1 (other)
 2
 6
 12 (herring)
 22 (saithe)
 30 (haddock)
 31 (cod)
}
