% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\name{plotConcentration}
\alias{plotConcentration}
\title{Plot concentration}
\usage{
plotConcentration(profile, header = "")
}
\arguments{
\item{profile}{\code{\link[acousticsQA]{verticalProfile}}}

\item{header}{header for plot}
}
\description{
Plots cumulative sa against cumulative distance,
 in order of decreasing sa
}
