% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\name{plotHorisontalProfile}
\alias{plotHorisontalProfile}
\title{Plots horisontal profile}
\usage{
plotHorisontalProfile(profile, header = "")
}
\arguments{
\item{profile}{\code{\link[acousticsQA]{horisontalProfile}}}

\item{header}{header for plot}
}
\description{
Plots aggregated sa binned by distance from bottom
}
\details{
Plots aggregated sa binned by distance from bottom
 Cumulative fractions are plotted on an alternate y-axis
}
