% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\name{plotLogNormalQQ}
\alias{plotLogNormalQQ}
\title{Plot lognormal QQ plot}
\usage{
plotLogNormalQQ(profile, header = "")
}
\arguments{
\item{profile}{\code{\link[acousticsQA]{verticalProfile}}, NAs are ignored.}

\item{header}{header for plot}
}
\description{
Plot quantiles of log-transformed sa against quantiles of normal distribution.
}
