% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\name{plotLogSaHistogram}
\alias{plotLogSaHistogram}
\title{Plot histogram of ln sa}
\usage{
plotLogSaHistogram(profile, header = "")
}
\arguments{
\item{profile}{\code{\link[acousticsQA]{verticalProfile}}, NAs are ignored.}

\item{header}{header for plot}
}
\description{
Plot distribution of ln sa over vertical bands as histogram
}
