% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\name{plotMap}
\alias{plotMap}
\title{Plot map}
\usage{
plotMap(
  profile,
  trawls = NULL,
  lonLim = NULL,
  latLim = NULL,
  projection = NULL,
  header = "",
  maxSaSize = 5,
  saColor = "black",
  trawlColor = "red",
  trawlPointShape = 10
)
}
\arguments{
\item{profile}{\code{\link[acousticsQA]{verticalProfile}} from acoustic registrations}

\item{trawls}{\code{\link[acousticsQA]{trawlLocation}}}

\item{lonLim}{limits for plots (longitudes). Vector of size 2. Will be calculated from 'profile' if NULL.}

\item{latLim}{limits for plots (latitudes). Vector of size 2. Will be calculated from 'profile' if NULL.}

\item{header}{header for plot}

\item{maxSaSize}{controls the maximum size of points representing vertical bands in 'profile'}

\item{saColor}{color of points representing vertical bands in 'profile'}

\item{trawlColor}{color of points representing trawl locations.}

\item{trawlPointShape}{code for point shape for trawl locations (ggplot2-codes)}
}
\description{
Plots map of acoustic registration, and optionally trawl positions.
 Vertical bands in acoustic registrations are represented as points with area proportional to registration (sa).
}
