% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\name{plotPurity}
\alias{plotPurity}
\title{Plot purity}
\usage{
plotPurity(
  profileTarget,
  profileAll,
  header = "",
  breaks = seq(0.5, 1, 0.05),
  ylim = c(0, 1)
)
}
\arguments{
\item{profileTarget}{\code{\link[acousticsQA]{verticalProfile}} from acoustic registrations of target acoustic category}

\item{profileAll}{\code{\link[acousticsQA]{verticalProfile}} from acoustic registrations of all acoustic categories}

\item{header}{header for plot}

\item{breaks}{vector of purities ([0,1]) to plot contributions for}

\item{ylim}{limits for y-axis.}
}
\description{
Plots contributions to total echo by the purity of target acoustic category.
}
\details{
Each interpreted segment is assigned a purity, which is the fraction of the total
 echo assigned to the acoustic category of interest (target acoustic category).
 The cumulative contribution to the total echo of the target acoustic category is plotted against
 different thresholds for purity. In the example provided below, one can for example infer that between
 80\% and 90\% of the total echo assigned to cod came from segments where at least 80 % of the echo was
 assigned to COD.

 Purity of assignment in a segment may serve as a proxy for confidence in the assignment,
 it typically reflects that categories are well separable acoustically, and that trawl-samples supports
 a relatively pure assignment.
}
\examples{
 prof <- verticalProfileLUF20(acousticsQA::echosounderSkrei2019, 31)
 profAll <- verticalProfileLUF20(acousticsQA::echosounderSkrei2019,
     unique(acousticsQA::echosounderSkrei2019$acocat$acocat))
 plotPurity(prof, profAll, header="COD")
}
