% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\name{plotStretch}
\alias{plotStretch}
\title{Plot vertical profile}
\usage{
plotStretch(profile, header = "")
}
\arguments{
\item{profile}{\code{\link[acousticsQA]{verticalProfile}}}

\item{header}{header for plot}
}
\description{
Plots aggregated sa for each vertical band
}
\details{
Plots aggregated sa for each vertical band
 Cumulative fractions are plotted on an alternate y-axis
}
