% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\name{trawlLocation}
\alias{trawlLocation}
\title{Trawl locations}
\description{
Trawl locations
}
\details{
\code{\link[data.table]{data.table}} with columns
 \describe{
  \item{log}{log, sailed distance in nmi}
  \item{time}{time as POSIXct}
  \item{latitude}{latitude in decimal degrees, WGS84}
  \item{longitude}{longitude in decimal degrees, WGS84}
 }
}
