% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\name{verticalProfileLUF20}
\alias{verticalProfileLUF20}
\title{Vertical profile}
\usage{
verticalProfileLUF20(
  echosounder,
  targetAcocat,
  channelType = "P",
  freq = 38000
)
}
\arguments{
\item{echosounder}{luf20 as read by RstoxData::readXmlFile}

\item{targetAcocat}{acoustic category to extract profile for}

\item{channelType}{the channel type to extract profile from (e.g. Pelagic (P) or Bottom (B))}

\item{freq}{frequncy of transreceiver to extract data from.}
}
\value{
\code{\link[acousticsQA]{verticalProfile}}
}
\description{
Makes vertical profile (sa by sailed distance)
}
\details{
Consult reference table for acoustic categories.
 https://referenceeditor.hi.no/apps/referenceeditor/v2/tables/acousticCategory

 some common options:
 1 (other)
 2
 6
 12 (herring)
 22 (saithe)
 30 (haddock)
 31 (cod)
}
